+++ Classes/ContentElement/TextContentElement.php
@@ -14,8 +14,9 @@ class TextContentElement extends AbstractContentElement
     {
         return [
             'text' => $this->getRteService()->parse(
-                $contentElementRawData['bodytext'],
-                Configurator::getByPid((int)$contentElementRawData['pid'])
+                (string)$contentElementRawData['bodytext'],
+                Configurator::getByPid((int)$contentElementRawData['pid']),
+                (int)$contentElementRawData['sys_language_uid']
             ),
         ];
     }