+++ Classes/Indexer/RecordIndexer.php
@@ -3,14 +3,14 @@
 namespace SourceBroker\Hugo\Indexer;
 
 use SourceBroker\Hugo\Configuration\Configurator;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Core\Database\ConnectionPool;
+
 use SourceBroker\Hugo\Domain\Model\DocumentCollection;
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
 use TYPO3\CMS\Extbase\DomainObject\AbstractDomainObject;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Extbase\Persistence\Generic\LazyObjectStorage;
-use TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 
 class RecordIndexer extends AbstractIndexer
@@ -32,7 +32,7 @@ class RecordIndexer extends AbstractIndexer
                     /** @var \TYPO3\CMS\Extbase\Persistence\Repository $repository */
                     $repository = $this->getRepositoryByTable($table);
 
-                    if(is_object($repository)) {
+                    if (is_object($repository)) {
                         $query = $repository->createQuery();
                         $query->getQuerySettings()->setRespectStoragePage(false);
 
@@ -48,7 +48,7 @@ class RecordIndexer extends AbstractIndexer
                     }
 
                     foreach ($recordRows as $record) {
-                        if(is_object($record) && $record instanceof AbstractDomainObject){
+                        if (is_object($record) && $record instanceof AbstractDomainObject) {
                             $record = $this->mapPropertiesToArrayRecursive($record->_getProperties());
                         }
                         $slug = $this->slugify($record['title']);
@@ -76,20 +76,20 @@ class RecordIndexer extends AbstractIndexer
 
         $repositoryClass = null;
         $modelSubclasses = [];
-        foreach($frameworkConfiguration['persistence']['classes'] as $class => $classConfig){
-            if($tableName == $dataMapper->convertClassNameToTableName($class)){
-                $tempRepositoryClass = str_replace('Model', 'Repository', $class).'Repository';
+        foreach ($frameworkConfiguration['persistence']['classes'] as $class => $classConfig) {
+            if ($tableName == $dataMapper->convertClassNameToTableName($class)) {
+                $tempRepositoryClass = str_replace('Model', 'Repository', $class) . 'Repository';
                 if (class_exists($tempRepositoryClass) && !in_array($class, $modelSubclasses)) {
                     $repositoryClass = $tempRepositoryClass;
 
-                    if($classConfig['subclasses']){
+                    if ($classConfig['subclasses']) {
                         $modelSubclasses = array_merge($modelSubclasses, $classConfig['subclasses']);
                     }
                 }
             }
         }
 
-        if(!empty($repositoryClass)) {
+        if (!empty($repositoryClass)) {
             return $objectManager->get($repositoryClass);
         }
 
@@ -98,21 +98,22 @@ class RecordIndexer extends AbstractIndexer
 
     protected function mapPropertiesToArrayRecursive(array $properties)
     {
-        foreach($properties as $property => $value)
-        {
-            if(is_object($value)) {
+        foreach ($properties as $property => $value) {
+            if (is_object($value)) {
 
-                if($value instanceof LazyObjectStorage || $value instanceof ObjectStorage){
+                if ($value instanceof LazyObjectStorage || $value instanceof ObjectStorage) {
                     $properties[$property] = $this->mapPropertiesToArrayRecursive($value->toArray());
-                } else if($value instanceof AbstractDomainObject){
-                    $properties[$property] = $this->mapPropertiesToArrayRecursive($value->_getProperties());
                 } else {
-                    $properties[$property] = $value;
+                    if ($value instanceof AbstractDomainObject) {
+                        $properties[$property] = $this->mapPropertiesToArrayRecursive($value->_getProperties());
+                    } else {
+                        $properties[$property] = $value;
+                    }
                 }
 
             }
 
-            if(empty($properties[$property])){
+            if (empty($properties[$property])) {
                 $properties[$property] = '';
             }
         }