+++ Classes/Task/ExportTask.php
@@ -17,6 +17,8 @@ class ExportTask extends \TYPO3\CMS\Scheduler\Task\AbstractTask
 
     /**
      * @return bool
+     * @throws \TYPO3\CMS\Core\Locking\Exception\LockAcquireException
+     * @throws \TYPO3\CMS\Core\Locking\Exception\LockCreateException
      */
     public function execute()
     {