+++ Classes/Tca/Pages.php
@@ -12,15 +12,18 @@ class Pages
 {
     /**
      * Returns Hugo Front Matter for given page
+     * @param $data
+     * @return string
      */
     public function getHugoFrontMatter($data)
     {
+        $content = '';
         if (!empty($data['row']['uid'])) {
             $configurator = Configurator::getByPid((int)$data['row']['uid']);
 
             $sysLanguageUid = 0; // TODO - do support for multilang
             $slugifiedRootline = GeneralUtility::makeInstance(\SourceBroker\Hugo\Utility\RootlineUtility::class,
-                $data['row']['uid'], $sysLanguageUid)->getSlugifiedRootlineForFilePath();
+                $data['row']['uid'], $sysLanguageUid)->getSlugifiedRootlinePath();
 
             $filename = PATH_site . rtrim($configurator->getOption('writer.path.content'), '/') .
                 '/' . $slugifiedRootline . '_index.md';
@@ -36,8 +39,8 @@ class Pages
                     . '<li>one of parent page is hidden,</li> '
                     . '<li>TYPO3 Hugo export failed.</li>'
                     . '</ul>';
-                }
-            return $content;
+            }
         }
+        return $content;
     }
 }