+++ Classes/Typolink/AbstractTypolinkBuilder.php
@@ -1,5 +1,6 @@
 <?php
-declare(strict_types = 1);
+declare(strict_types=1);
+
 namespace SourceBroker\Hugo\Typolink;
 
 /*
@@ -16,7 +17,6 @@ namespace SourceBroker\Hugo\Typolink;
  */
 
 use SourceBroker\Hugo\Configuration\Configurator;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
@@ -32,18 +32,25 @@ abstract class AbstractTypolinkBuilder extends \TYPO3\CMS\Frontend\Typolink\Abst
     protected $txHugoConfigurator;
 
     /**
+     * @var int
+     */
+    protected $txHugoSysLanguageUid = 0;
+
+    /**
      * AbstractTypolinkBuilder constructor.
      *
      * @param ContentObjectRenderer $contentObjectRenderer
      * @param Configurator $txHugoConfigurator
+     * @param int $txHugoSysLanguageUid
      */
     public function __construct(
         ContentObjectRenderer $contentObjectRenderer,
-        Configurator $txHugoConfigurator
-    )
-    {
+        Configurator $txHugoConfigurator,
+        int $txHugoSysLanguageUid = 0
+    ) {
         parent::__construct($contentObjectRenderer);
         $this->txHugoConfigurator = $txHugoConfigurator;
+        $this->txHugoSysLanguageUid = $txHugoSysLanguageUid;
     }
 
     /**
@@ -63,8 +70,12 @@ abstract class AbstractTypolinkBuilder extends \TYPO3\CMS\Frontend\Typolink\Abst
      *
      * {@inheritdoc}
      */
-    protected function resolveTargetAttribute(array $conf, string $name, bool $respectFrameSetOption = false, string $fallbackTarget = ''): string
-    {
+    protected function resolveTargetAttribute(
+        array $conf,
+        string $name,
+        bool $respectFrameSetOption = false,
+        string $fallbackTarget = ''
+    ): string {
         if (isset($conf[$name])) {
             $target = $conf[$name];
         } else {
@@ -73,48 +84,56 @@ abstract class AbstractTypolinkBuilder extends \TYPO3\CMS\Frontend\Typolink\Abst
         if ($conf[$name . '.']) {
             $target = (string)$this->contentObjectRenderer->stdWrap($target, $conf[$name . '.']);
         }
+
         return $target;
     }
 
     /**
-     * Overwrites method to get the link target to not use TSFE inside of it
-     *
-     * {@inheritdoc}
+     * @return callable[]
      */
-    protected function forceAbsoluteUrl(string $url, array $configuration): string
+    protected function getProcessors(): array
     {
-        if (!empty($url) && !empty($configuration['forceAbsoluteUrl']) &&  preg_match('#^(?:([a-z]+)(://)([^/]*)/?)?(.*)$#', $url, $matches)) {
-            $urlParts = [
-                'scheme' => $matches[1],
-                'delimiter' => '://',
-                'host' => $matches[3],
-                'path' => $matches[4]
-            ];
-            $isUrlModified = false;
-            // Set scheme and host if not yet part of the URL:
-            if (empty($urlParts['host'])) {
-                $urlParts['scheme'] = GeneralUtility::getIndpEnv('TYPO3_SSL') ? 'https' : 'http';
-                $urlParts['host'] = GeneralUtility::getIndpEnv('HTTP_HOST');
-                $urlParts['path'] = '/' . ltrim($urlParts['path'], '/');
-                // absRefPrefix has been prepended to $url beforehand
-                // so we only modify the path if no absRefPrefix has been set
-                // otherwise we would destroy the path
-                if ($this->txHugoConfigurator->getOption('link.absRefPrefix') === '') {
-                    $urlParts['path'] = GeneralUtility::getIndpEnv('TYPO3_SITE_PATH') . ltrim($urlParts['path'], '/');
-                }
-                $isUrlModified = true;
-            }
-            // Override scheme:
-            $forceAbsoluteUrl = &$configuration['forceAbsoluteUrl.']['scheme'];
-            if (!empty($forceAbsoluteUrl) && $urlParts['scheme'] !== $forceAbsoluteUrl) {
-                $urlParts['scheme'] = $forceAbsoluteUrl;
-                $isUrlModified = true;
-            }
-            // Recreate the absolute URL:
-            if ($isUrlModified) {
-                $url = implode('', $urlParts);
-            }
+        return [];
+    }
+
+    /**
+     * @param string $url
+     *
+     * @return string
+     */
+    protected function applyHugoProcessors(string $url): string
+    {
+        /** @var callable $processor */
+        foreach ($this->getProcessors() as $processor) {
+            $url = $processor($url);
         }
+
         return $url;
+    }
+
+    /**
+     * @param string $url
+     *
+     * @return string
+     */
+    protected function addHugoAbsRelPrefix(string $url): string
+    {
+        return $this->txHugoConfigurator->getOption('link.absRefPrefix') . $url;
+    }
+
+    /**
+     * @param string $url
+     *
+     * @return string
+     */
+    protected function addHugoLanguagePrefix(string $url): string
+    {
+        $langPrefix = $this->txHugoConfigurator->getOption('languages.' . $this->txHugoSysLanguageUid);
+
+        if (empty($this->txHugoSysLanguageUid) || empty($langPrefix)) {
+            return $url;
+        }
+
+        return $langPrefix . '/' . $url;
     }
 }