+++ Classes/Typolink/PageLinkBuilder.php
@@ -19,7 +19,6 @@ namespace SourceBroker\Hugo\Typolink;
 use SourceBroker\Hugo\Domain\Repository\Typo3PageRepository;
 use SourceBroker\Hugo\Utility\RootlineUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Core\Utility\MathUtility;
 
 /**
  * Builds a TypoLink to a certain page
@@ -32,13 +31,43 @@ class PageLinkBuilder extends AbstractTypolinkBuilder
     public function build(array &$linkData, string $linkText, string $target, array $conf): array
     {
         $url = null;
-        $pageUid = $linkData['pageuid'];
-        if (MathUtility::canBeInterpretedAsInteger($pageUid)) {
-            $page = GeneralUtility::makeInstance(Typo3PageRepository::class)->getByUid((int)$pageUid);
+
+        // support for some legacy links to the current page
+        // @see \TYPO3\CMS\Core\LinkHandling\LegacyLinkNotationConverter::resolvePageRelatedParameters
+        $pageUid = ($linkData['pageuid'] !== 'current')
+            ? (int)$linkData['pageuid']
+            : $this->txHugoConfigurator->getPageUid()
+        ;
+
+        if ($pageUid) {
+            $page = GeneralUtility::makeInstance(Typo3PageRepository::class)->getByUid($pageUid);
             if ($page['hidden'] === 0 && $page['deleted'] === 0) {
-                $url = GeneralUtility::makeInstance(RootlineUtility::class, $pageUid)->getSlugifiedRootlineForUrl();
+                $url = GeneralUtility::makeInstance(RootlineUtility::class, $pageUid)
+                    ->getSlugifiedRootlinePath($this->txHugoSysLanguageUid);
+
+                if (isset($linkData['fragment'])) {
+                    $url .= '#' . $linkData['fragment'];
+                }
             }
         }
-        return [$url, empty($linkText) ? $page['title'] : $linkText, $target];
+        return [
+            $url === null ? null : $this->applyHugoProcessors($url),
+            (empty($linkText) && !empty($page['title'])) ? $page['title'] : $linkText,
+            $target
+        ];
+    }
+
+    /**
+     * @return callable[]
+     */
+    protected function getProcessors(): array
+    {
+        return array_merge(
+            parent::getProcessors(),
+            [
+                [$this, 'addHugoLanguagePrefix'],
+                [$this, 'addHugoAbsRelPrefix'],
+            ]
+        );
     }
 }