+++ ext_tables.sql
@@ -2,6 +2,11 @@ CREATE TABLE tx_dce_domain_model_dce (
 	tx_hugo_typename varchar(255),
 );
 
+CREATE TABLE sys_domain (
+	tx_hugo_domains varchar(255),
+);
+
+
 CREATE TABLE queue_items (
   uid int(11) NOT NULL auto_increment,
   namespace varchar (32) NOT NULL,