+++ CHANGELOG.rst
@@ -1,6 +1,34 @@
 Changelog
 ---------
 
+master
+~~~~~~
+1) [BUGFIX] Do not break on non int values in link.
+2) [TASK] Remove support for legacy links.
+3) [BUGFIX] Bugfix for edge cases with link creation.
+4) [FEATURE] Add Hugo page view buttons into page context-sensitive menu.
+5) [FEATURE] Implement support of content slide functionality.
+6) [TASK] Implement support for links with fragments and legacy links to the current page.
+
+0.5.0
+~~~~~
+1) [FEATURE] Add pid to content exported data.
+2) [BUGFIX] PageIndexer - resolve the problem with elements from parent pages.
+
+0.4.0
+~~~~~
+1) [TASK] Optimize getSlugifiedRootline / add two more methods for slugified rootline: getSlugifiedRootlineForUrl
+2) [TASK] Implement support for `link.absRefPrefix` into file, folder and page link builders
+3) [TASK] Cleanup - Remove testing line for preg_match
+4) [TASK] refactor service classes
+5) [TASK] Handle undelete of the content element action
+6) [BUGFIX] Fix indexing of translated content elements
+7) [TASK] refactor class Typo3PageRepository; PageIndexer - add the content elements from parent
+8) [FEATURE] Implement grid content element exporter
+9) [TASK] Implement multilang support into link builders
+10) [TASK] Code cleanup / PSR-2 / phpdocs
+11) [BUGFIX] Remove linkText form generated url make htmlspecialchars for title
+
 0.3.0
 ~~~~~
 1) [TASK] Optimize getSlugifiedRootline / add two more methods for slugified rootline: getSlugifiedRootlineForUrl